requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf
PyPDF2
fuzzywuzzy[speedup]
python-dotenv
